Moving Array class to bitstring_array.py.
